+++ src/main/java/org/opentripplanner/routing/api/request/preference/StreetPreferences.java
@@ -14,6 +14,7 @@ import org.opentripplanner.routing.api.request.framework.DurationForEnum;
 import org.opentripplanner.routing.api.request.framework.Units;
 import org.opentripplanner.routing.core.intersection_model.DrivingDirection;
 import org.opentripplanner.routing.core.intersection_model.IntersectionTraversalModel;
+import org.opentripplanner.standalone.config.routerequest.RouteRequestMapper;
 import org.opentripplanner.util.lang.DoubleUtils;
 import org.opentripplanner.util.lang.ToStringBuilder;
 
@@ -81,12 +82,12 @@ public final class StreetPreferences implements Serializable {
     return intersectionTraversalModel;
   }
 
-  /** see {@link org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper} **/
+  /** see {@link RouteRequestMapper} **/
   public DurationForEnum<StreetMode> maxAccessEgressDuration() {
     return maxAccessEgressDuration;
   }
 
-  /** see {@link org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper} **/
+  /** see {@link RouteRequestMapper} **/
   public DurationForEnum<StreetMode> maxDirectDuration() {
     return maxDirectDuration;
   }
