+++ src/main/java/org/opentripplanner/routing/api/request/preference/SystemPreferences.java
@@ -10,6 +10,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 import org.opentripplanner.ext.dataoverlay.api.DataOverlayParameters;
 import org.opentripplanner.routing.api.request.RoutingTag;
+import org.opentripplanner.standalone.config.routerequest.RouteRequestMapper;
 import org.opentripplanner.util.lang.ToStringBuilder;
 
 /**
@@ -60,7 +61,7 @@ public class SystemPreferences implements Serializable {
     return tags;
   }
 
-  /** see {@link org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper} **/
+  /** see {@link RouteRequestMapper} **/
   public DataOverlayParameters dataOverlay() {
     return dataOverlay;
   }
@@ -70,7 +71,7 @@ public class SystemPreferences implements Serializable {
     return geoidElevation;
   }
 
-  /** see {@link org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper} **/
+  /** see {@link RouteRequestMapper} **/
   public Duration maxJourneyDuration() {
     return maxJourneyDuration;
   }
