+++ src/main/java/org/opentripplanner/routing/api/request/preference/WalkPreferences.java
@@ -6,6 +6,7 @@ import java.io.Serializable;
 import java.util.Objects;
 import java.util.function.Consumer;
 import org.opentripplanner.routing.api.request.framework.Units;
+import org.opentripplanner.standalone.config.routerequest.RouteRequestMapper;
 import org.opentripplanner.util.lang.ToStringBuilder;
 
 /**
@@ -89,7 +90,7 @@ public final class WalkPreferences implements Serializable {
     return stairsReluctance;
   }
 
-  /** see {@link org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper} **/
+  /** see {@link RouteRequestMapper} **/
   public double stairsTimeFactor() {
     return stairsTimeFactor;
   }
