+++ src/main/java/org/opentripplanner/standalone/config/buildconfig/TransferRequestConfig.java
@@ -5,7 +5,7 @@ import static org.opentripplanner.standalone.config.framework.json.OtpVersion.NA
 import java.util.List;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
-import org.opentripplanner.standalone.config.routerequest.RoutingRequestMapper;
+import org.opentripplanner.standalone.config.routerequest.RouteRequestMapper;
 
 public class TransferRequestConfig {
 
@@ -15,6 +15,6 @@ public class TransferRequestConfig {
       .since(NA)
       .summary("TODO")
       .description(/*TODO DOC*/"TODO")
-      .asObjects(List.of(new RouteRequest()), RoutingRequestMapper::mapRouteRequest);
+      .asObjects(List.of(new RouteRequest()), RouteRequestMapper::mapRouteRequest);
   }
 }
